+++ engine/src/ui/ui.c
@@ -10,16 +10,10 @@ void UIRender(UI* ui, void* context)
 {
     castor_Context* ctx = context;
 
-    // TOASTS
-    for(size_t i = 0; i < ui->toast_count; i++){
-        UIToast toast = ui->toasts[i];
-        if (toast.active) {
-            Uint32 elapsed_time = SDL_GetTicks() - toast.start_time;
-            if (elapsed_time < toast.duration) {
-                UIToastRender(ctx->sdl.renderer, ui->font, &toast, ctx->sdl.screen_width, ctx->sdl.screen_height);
-            } else {
-                toast.active = false;
-            }
+    for (size_t i = 0; i < ui->toast_count; i++) {
+        UIToast* toast = &ui->toasts[i];
+        if (toast->active) {
+            UIToastRender(ctx->sdl.renderer, ui->font, toast, ctx->sdl.screen_width, ctx->sdl.screen_height);
         }
     }
 }